Move youtube downloader to a package on root level
