+++ app/src/main/java/com/yurii/youtubemusic/screens/youtube/playlists/Playlist.kt
@@ -1,4 +1,4 @@
-package com.yurii.youtubemusic.screens.youtube.models
+package com.yurii.youtubemusic.screens.youtube.playlists
 
 import com.google.api.services.youtube.model.Playlist as YouTubePlaylistModel
