+++ app/src/main/java/com/yurii/youtubemusic/screens/youtube/playlists/PlaylistsAdapter.kt
@@ -8,7 +8,6 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.yurii.youtubemusic.R
 import com.yurii.youtubemusic.databinding.ItemPlaylistBinding
-import com.yurii.youtubemusic.screens.youtube.models.Playlist
 
 class PlaylistsAdapter(private val selectedPlaylist: Playlist?, private val onSelectedPlaylist: (Playlist) -> Unit) :
     PagingDataAdapter<Playlist, PlaylistsAdapter.PlaylistViewHolder>(Comparator) {
